Add the latest tag to images built from main.
I thought this would happen automatically, but it turns out by default
it is only done for Git tags. See the [`docker/metadata-action`][metadata-action]
docs for details on this.

The rest of the tags spec is there to retain the default behaviour of
creating tags based on branch names and PR numbers.

[metadata-action]: https://github.com/docker/metadata-action/blob/b798ed8388581f6f002541048b6458ca2c4ba442/README.md#latest-tag